pyproject.toml
[project]
name = "customerassistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "beautifulsoup4>=4.13.4",
    "chromadb>=1.0.15",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "ipython>=9.4.0",
    "jinja2>=3.1.6",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.17",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.6",
    "langchain-openai>=0.3.28",
    "langgraph>=0.5.4",
    "markdown>=3.8.2",
    "openpyxl>=3.1.5",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pydantic>=2.11.7",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "sqlalchemy>=2.0.41",
]
